Refactor authentication pages: - Migrate SignUp component to a separate file and simplify the signup page structure. - Update SignIn component with similar structure and functionality. - Change navigation links for forgot password and reset password to point to the signin page. - Update job seeker and recruiter sections to link to the signup page. - Modify Next.js configuration to use remote patterns for image domains.
