+++ components/jobSeeker/Section.js
@@ -46,7 +46,7 @@ function Section1() {
         </p>
 
         <Link
-          href={"/auth"}
+          href={"/auth/signup"}
           className="bg-primary text-white text-xl leading-6 font-medium py-4 px-8 rounded-lg mt-12"
         >
           Sign me up!
